* fix root dir

+++ scripts/apollo_base.sh
@@ -447,7 +447,9 @@ fi
 check_in_docker
 create_data_dir
 
-sudo ln -s $APOLLO_ROOT_DIR /apollo
+if [ ! -e /apollo ];then
+  sudo ln -s $APOLLO_ROOT_DIR /apollo
+fi
 
 if [ -z $APOLLO_BASE_SOURCED ]; then
   set_lib_path
